refactor(app): break the app/pane/tab/bands dependency cycle
`pane` and `bands` imported `PlotAreas`, `fmt_time_as`, `plot_split`,
`split_time_strip` and `gesture_hits_lane_divider` from `app` — the
composition root that imports them back. With `tab` sitting between,
the four modules formed the crate's only strongly connected component,
so the repository's two largest files could not move without each other.

Move the geometry and the axis labels down into a new `plot_area`, below
all three: given a rect it returns rects, given an instant it returns a
string, and it reads no application state. `app` now imports `pane` and
`tab` in one direction only, and nothing imports `app` back.

The tests covering the moved items travel with them — four from
`panes_layout_tests`, one from `drawings_tests`, two from
`chart_view_tests` — into `plot_area`'s own test module. No signature
changes: every call site differs only in the module path.
