+++ crates/app/src/app/tests/drawings_tests.rs
@@ -1,26 +1,5 @@
 use super::*;
 
-/// The divider is the tape's, and nothing else in the band is.
-///
-/// The predicate this replaced claimed the *whole* lane, so a third of the
-/// canvas took a press and did nothing with it: the trader pulled and the
-/// chart did not move, with nothing on screen saying why. The wheel had
-/// already been handed back to the candles for exactly that reason; this is
-/// the drag catching up.
-#[test]
-fn only_the_divider_handle_is_off_limits_to_the_pan() {
-    let half = 5.0;
-    assert!(gesture_hits_lane_divider(Some(700.0), 700.0, half));
-    assert!(gesture_hits_lane_divider(Some(700.0), 695.0, half));
-    assert!(gesture_hits_lane_divider(Some(700.0), 705.0, half));
-    // The band beyond the handle belongs to the candles again.
-    assert!(!gesture_hits_lane_divider(Some(700.0), 706.0, half));
-    assert!(!gesture_hits_lane_divider(Some(700.0), 1_200.0, half));
-    assert!(!gesture_hits_lane_divider(Some(700.0), 694.0, half));
-    // No lane, no handle.
-    assert!(!gesture_hits_lane_divider(None, 700.0, half));
-}
-
 #[test]
 fn a_quote_driven_feed_says_so_where_the_side_note_goes() {
     let (evt_tx, evt_rx) = mpsc::channel(64);
